requirements.txt
pygame-ce
python-dotenv